fix(quote): StockDetailItem.Etf type int? → double?
stock_detail returns etf as a float (0.0 / 1.0), triggering
  'Input string 0.0 is not a valid integer. Path data.items[0].etf'
on deserialization. Widen to double? to tolerate the wire form.

Third of three post-un-skip type/field fixes for stock_detail;
previous two were latestTime (string?) and Fields enum values.

+++ TigerOpenAPI/Quote/Response/StockDetailItem.cs
@@ -118,8 +118,12 @@ namespace TigerOpenAPI.Quote.Response
     /// signalling non-leveraged / 2x / 3x, and negative values marking
     /// inverse ETFs (matches Python doc).
     /// </summary>
+    /// <summary>
+    /// stock_detail returns etf as a float (0.0 / 1.0), not an integer flag.
+    /// Use double? to tolerate both forms.
+    /// </summary>
     [JsonProperty(PropertyName = "etf")]
-    public int? Etf { get; set; }
+    public double? Etf { get; set; }
 
     /// <summary>Listing date (epoch millis at 00:00 in the market's local tz).</summary>
     [JsonProperty(PropertyName = "listingDate")]
